requirements.txt
flask==3.1.0
pandas==2.2.3
scikit-learn==1.6.1
joblib==1.4.2
numpy==2.2.3
requests==2.31.0
python-dotenv==1.0.1
gunicorn==21.2.0
prometheus-client==0.20.0
pytest==8.0.2
black==24.2.0
flake8==7.0.0 